refactor: WidgetGrid로 추출
